Encrypt and Decrypt tool, and load remote config file
