+++ src/main/java/com/alibaba/druid/filter/config/security/tool/Action.java
@@ -1,0 +1,11 @@
+package com.alibaba.druid.filter.config.security.tool;
+
+/**
+ * @author Jonas Yang
+ */
+public interface Action {
+
+    public String getId();
+
+    public void execute();
+}
