+++ src/main/resources/META-INF/druid-config-loader-default.properties
@@ -1,2 +1,2 @@
-config.loader.1=com.alibaba.druid.filter.config.impl.FileConfigLoader
-config.loader.2=com.alibaba.druid.filter.config.impl.HttpConfigLoader
+config.loader.1=com.alibaba.druid.filter.config.loader.impl.FileConfigLoader
+config.loader.2=com.alibaba.druid.filter.config.loader.impl.HttpConfigLoader
